WIP ch06 6-04 freq jankunit FORK proof of concept
POC_stdout_stderr_multipipe_separation.c

This file presents a proof of concept of separating stdout/stderr
of different parts of a child process, whereas the stdout/stderr
of different code blocks are redirected to different sets of pipes
based on conditions. The pipes will then be read in the parent for
processing

This will be used by the hypothetical FORK macro of the jankunit
unit testing library, whereby users will be able to test functions
in their codebases that can exit/signal or print to stdout/stderr.

All the while, the assertion and exception checking results within
the forked block of code (i.e. code running in the child) will
still be able to be relayed to the parent for proper formatting
and printing to the parent stdout.