ADD ch06 6-04 freq freq.internal.h freq.internal.c
Added ea_readline function
    reads file descriptor line by line, allocates a buf,
buf and relevant metadata need to be maintained by the caller across calls.

    Calling readline again with same arguments will return the same line.

    First call needs to be made with the address of a NULL *buf in order to
read the file descriptor, else the buffer will be tried to be read and the
nbuf_allocd needs to be the same as the size of non-null *buf.

    returns the length of line read, assigns char **ln
to the beginning of the line found in buf, returns -1 for EOF

    Example:
        char *buf = NULL;
        int nbuf_allocd;
        char *ln;
        int buf_seek = 0;

        int len1 = ea_readline(fd, &buf, &nbuf_allocd, &ln, buf_seek);
        printf("%.*s\n", len1, ln);

        buf_seek = len1 + 1;
        int len2 = ea_readline(fd, &buf, &nbuf_allocd, &ln, buf_seek);
        printf("%.*s\n", len2, ln);

        ...

Added ea_read_buf function
    reads given file descriptor, allocates memory for a buf that is maintained
by the caller. Null terminates the buf when input fd is exhausted.

    returns the number of bytes allocated, exits with message if error

    Example:
        int fd;
        char *buf;
        int allocated = ea_read_buf(fd, &buf);